# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/category/add" : {
      "post" : {
        "summary" : "add",
        "description" : "add",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "categoryAlias", "categoryName" ],
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : "主键ID"
                  },
                  "categoryName" : {
                    "type" : "string",
                    "description" : "分类名称"
                  },
                  "categoryAlias" : {
                    "type" : "string",
                    "description" : "分类别名"
                  },
                  "createUser" : {
                    "type" : "integer",
                    "description" : "创建人ID"
                  },
                  "createTime" : {
                    "type" : "string",
                    "description" : "创建时间"
                  },
                  "updateTime" : {
                    "type" : "string",
                    "description" : "更新时间"
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/category/detail" : {
      "get" : {
        "summary" : "detail",
        "description" : "detail",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "array",
                      "description" : "响应数据",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "id" : {
                            "type" : "integer",
                            "description" : "主键ID"
                          },
                          "categoryName" : {
                            "type" : "string",
                            "description" : "分类名称"
                          },
                          "categoryAlias" : {
                            "type" : "string",
                            "description" : "分类别名"
                          },
                          "createUser" : {
                            "type" : "integer",
                            "description" : "创建人ID"
                          },
                          "createTime" : {
                            "type" : "string",
                            "description" : "创建时间"
                          },
                          "updateTime" : {
                            "type" : "string",
                            "description" : "更新时间"
                          }
                        },
                        "description" : "com.heima.big_event.pojo.Category"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/category/update" : {
      "put" : {
        "summary" : "update",
        "description" : "update",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "categoryAlias", "categoryName", "id" ],
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : "主键ID"
                  },
                  "categoryName" : {
                    "type" : "string",
                    "description" : "分类名称"
                  },
                  "categoryAlias" : {
                    "type" : "string",
                    "description" : "分类别名"
                  },
                  "createUser" : {
                    "type" : "integer",
                    "description" : "创建人ID"
                  },
                  "createTime" : {
                    "type" : "string",
                    "description" : "创建时间"
                  },
                  "updateTime" : {
                    "type" : "string",
                    "description" : "更新时间"
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/category/delete" : {
      "delete" : {
        "summary" : "delete",
        "description" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/add" : {
      "post" : {
        "summary" : "addCommentLike",
        "description" : "addCommentLike",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        }, {
          "name" : "commentId",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}